+++ src/main/java/chess/domain/piece/moving/KnightMoving.java
@@ -1,9 +1,9 @@
 package chess.domain.piece.moving;
 
 import chess.domain.Position;
+import chess.domain.game.ImpossibleMoveException;
 import chess.domain.piece.direction.KnightDirections;
 import chess.domain.piece.direction.PieceDirections;
-import chess.exception.ImpossibleMoveException;
 
 import java.util.ArrayList;
 import java.util.List;
