+++ src/main/java/chess/domain/piece/moving/PawnMoving.java
@@ -1,8 +1,8 @@
 package chess.domain.piece.moving;
 
 import chess.domain.Position;
+import chess.domain.game.ImpossibleMoveException;
 import chess.domain.piece.direction.PawnDirections;
-import chess.exception.ImpossibleMoveException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,8 +25,7 @@ public class PawnMoving implements PieceMoving {
     public void updateMovablePositions(List<Position> existPiecePositions, List<Position> enemiesPositions) {
         movablePositions = directions.movablePositions(currentPosition, existPiecePositions, enemiesPositions);
         if (notMoved()) {
-            directions.additionalMovePosition(currentPosition, existPiecePositions)
-                    .ifPresent(position -> movablePositions.add(position));
+            directions.additionalMovePosition(currentPosition, existPiecePositions, movablePositions);
         }
     }
